requirements.txt
unstructured
tiktoken
pinecone
pypdf
openai
sentence_transformers
langchain-community
pandas
numpy
python-dotenv


fastapi
uvicorn
langchain
sentence-transformers
pinecone-client>=3.0.0
python-dotenv